pyproject.toml
[project]
name = "tbc"
author = "qzdn"
version = "0.1.0"
description = "Simple API for Twitch bot commands"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastapi[standard]==0.119.*",
    "howlongtobeatpy==1.0.*",
    "python-dotenv==1.1.*",
]
